Refactor check to work around weird case with BSB/Generals like the Slann
